qa-lab: share Docker runtime helpers
